+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 LATEST CHANGES
 ==============
 
+2025-10-09
+----------
+- New feature: 3D visualization of the positional tracking landmarks as a point cloud on topic `~/pose/landmarks`
+  - Added parameter `pos_tracking.publish_3d_landmarks` to enable/disable landmarks publishing
+  - Added parameter `pos_tracking.publish_lm_skip_frame` to set the frequency of landmarks publishing (0 to publish every frame)
+
 2025-09-30
 ----------
 - Enabled Isaac ROS NITROS integration for ZED X One cameras